refactor: separate sixel tests
